// app/lib/ledger/types/AdminDashboardReport.ts
/**
 * This file is auto-generated. Do not edit manually.
 * Generated from Kotlin data classes.
 */

import { UserCaloriesAverage } from "./UserCaloriesAverage";
import { WeeklyComparisonReport } from "./WeeklyComparisonReport";

export type AdminDashboardReport = {
  weeklyComparison: WeeklyComparisonReport;
  userAverages: UserCaloriesAverage[];
};